Merged differnt people's work together.
Scripts now work together and have been added to a scene.
Unit class now allows for units that don't move.
Added the Unit's team to the unit.
Improvements to Turn Manager.
Health Bar now checks to see if the unit is dead when it takes damage instead of during Update()